feat: разнесла код на разные модули
