+++ js/constants.js
@@ -1,0 +1,43 @@
+const ADS_AMOUNT = 10;
+const TITLES = [
+  'Сдается уютная квартирка недалеко от метро',
+  'Сдам квартиру. Есть все условия для комфортного проживания',
+  'Лучшая квартира на районе!',
+  'Уютная квартира в центре города',
+  'Сдается квартира посуточно',
+  'Уютный причал для уставшего путника',
+];
+const DESCRIPTIONS = [
+  'Просторная квартира в туристическом районе, недалеко от метро, есть все необходимое для комфортного отдыха',
+  'Сдается посуточно квартира, со всеми условиями. Отличный вид с балкона',
+  'Лучше жилья вы не найдете! мы вам гарантируем!' +
+  'Японское гостеприимство в сочетании с европейским интерьером не оставят вас равнодушным! Вам позавидует сам Будда!',
+  'Лучшее предложение недалеко от Акихабары! Рядом находятся автобусные остановки, в квартире раздельный санузел,' +
+  'имеется комфортабельная кухня',
+];
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+const CHECKIN_AND_CHECKOUT_TIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+export { ADS_AMOUNT, TITLES, DESCRIPTIONS, TYPES, CHECKIN_AND_CHECKOUT_TIMES, FEATURES, PHOTOS };
